Migrate Mutation.deleteOrganization
